Fix raster dimensions and index access off-by-one of SRTMTileGrid
This would have resulted in an ArrayIndexOutOfBoundsException being thrown when attempting to access the northernmost or easternmost elevation values in an SRTMTileGrid. In practice, it did not appear, because, due to adjacent SRTM tiles overlapping by one row and column, we would create a new SRTMTileGrid instead of accessing elevation via the very last row or column of the current SRTMTileGrid.