Fix raster index calculation for bounds identical to grid bounds
This bug affected the largest integer latitude and longitude covered by the tile grid. It should not have caused problems in practice, because we would create a new tile grid with tiles to the north or east of the previous grid instead of accessing the very last column or row of data in a tile grid.